01_kafka/02_mqtt_to_kafka/requirements.txt
confluent-kafka==2.8.0
paho-mqtt==1.6.1
python-dotenv
pytz